requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-taggit==3.0.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
